Added navx-mxp (IMU) library code
This is for the navX-mxp inertial measurement unit board.  It provides
accelerometer, gyro and compass functionality.  This code is usually
installed to C:\Users\<user>\navx-mxp\cpp but we want it to be in our
repo.  Configure Eclipse to point to them inside the repo.  Generic
install details can be found here:
   http://www.pdocs.kauailabs.com/navx-mxp/software/roborio-libraries/c/