Moved the navx-mxp library outside the project
Figured out how to point to it outside the library.  Now we don't have
to see it (and exclude it from compile) in the robot code project.
Much cleaner.  Projects wanted to add the library need to add an include
to "${ProjDirPath}/../../../libs/navx-mxp/include" and a library path
to "${ProjDirPath}/../../../libs/navx-mxp/lib"