Added Camera Calibration Worksheet
Input Centre X, Centre Y and Tilt encoder values to get the coefficients
for the distance, rotate and tilt formulas.  The spreadsheet can hold
two trial to compare past and current values.  It calculates the best-
fit formula on the graph and in cell-format for simpler access to those
numbers.  It also includes a verification tab where you can input the
distance and see the expected Centre X, Centre Y and Tilt values for
comparison to the values received on the robot.